+++ catalyst/dl/__main__.py
@@ -2,9 +2,14 @@ from argparse import ArgumentParser, RawTextHelpFormatter
 from collections import OrderedDict
 
 from catalyst.__version__ import __version__
-from catalyst.dl.scripts import init, run, trace
+from catalyst.dl.scripts import run, trace
+from catalyst.tools.settings import IS_GIT_AVAILABLE
 
-COMMANDS = OrderedDict([("init", init), ("run", run), ("trace", trace)])
+COMMANDS = OrderedDict([("run", run), ("trace", trace)])
+if IS_GIT_AVAILABLE:
+    from catalyst.dl.scripts import init
+
+    COMMANDS["init"] = init
 
 
 def build_parser() -> ArgumentParser: